requirements/base.txt
# Base
pytz
python-slugify
Pillow
psycopg2 --no-binary psycopg2
typed-ast

# Django
django

# Environment
django-environ

# Django REST Framework
djangorestframework
django-filter

# PyJWT
PyJWT

# Passwords security
argon2-cffi

# Static files
whitenoise

# Celery
tornado
redis
django-redis
celery
flower
